requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-cors-headers==3.10.0
djangorestframework-simplejwt==5.0.0
Pillow==8.4.0